Wieder warnungsfrei
Elf Warnungen, alle mit dem Hall hereingekommen. Das Projekt hat den Anspruch,
bei voller JUCE-Warnschaerfe warnungsfrei zu bauen (siehe ARCHITEKTUR.md), also
sind sie weg statt stehen zu bleiben.

Neun davon sind Verdeckungen (-Wshadow): ein innerer Name deckt einen
gleichnamigen aeusseren zu. Das ist nie gewollt und im Zweifel ein Fehler -
wer "buffer" liest, meint den, den er drei Zeilen darueber angelegt hat.

  ReverbParts.h     der Seed wurde als int mit einer unsigned Konstante
                    multipliziert (-Wsign-conversion). Die Umwandlung steht
                    jetzt vor der Multiplikation, das Ergebnis ist dasselbe.
  OpenAirReverb.h   Rng-Konstruktorparameter "seed" deckte das Feld "seed" zu
  TapBus.h          t/l/r in der Spiegelschleife deckten die gleichnamigen
                    des geraden Weges zu, "panorama" in setMirrorReturn das
                    Feld gleichen Namens
  PluginEditor.cpp  Laufvariable "p" in der Elternkette
  PluginProcessor   "coeff" der Propellerrichtung deckte den des Gases zu
  load_check.cpp    buffer/midi in drei Szenarien
  scope_boom_probe  double auf float ohne Ansage

Kein Verhalten geaendert - es sind Umbenennungen und eine Umwandlung, die
vorher implizit an derselben Stelle passierte.

ctest: 5/5 gruen, Bau ohne eine einzige Warnung.

+++ Tests/scope_boom_probe.cpp
@@ -34,7 +34,7 @@ std::vector<float> makeSignal (double seconds, double boomEverySeconds, double q
     {
         seed = seed * 1664525u + 1013904223u;
 
-        x[(size_t) n] = quietLevel * (2.0f * ((float) (seed >> 8) / 16777216.0f) - 1.0f);
+        x[(size_t) n] = (float) quietLevel * (2.0f * ((float) (seed >> 8) / 16777216.0f) - 1.0f);
     }
 
     // Die Knalle selbst: eine kurze N-Welle, wie sie das Plugin erzeugt.
